Upgrade to arrow/parquet 55, and `object_store` to `0.12.0` and pyo3 to `0.24.0` (#15466)
* Temp pin to datafusion main

* Update cargo lock

* update pyo3

* vendor random generation

* Update error message

* Update for extraction

* Update pin

* Upgrade object_store

* fix feature

* Update file size handling

* bash for object store API changes

* few more

* Update APIs more

* update expected message

* update error messages

* Update to apache

* Update API for nicer parquet u64s

* Fix wasm build

* Remove pin

* Fix signature

+++ datafusion/core/src/datasource/physical_plan/parquet.rs
@@ -1786,13 +1786,13 @@ mod tests {
         path: &str,
         store: Arc<dyn ObjectStore>,
         batch: RecordBatch,
-    ) -> usize {
+    ) -> u64 {
         let mut writer =
             ArrowWriter::try_new(BytesMut::new().writer(), batch.schema(), None).unwrap();
         writer.write(&batch).unwrap();
         writer.flush().unwrap();
         let bytes = writer.into_inner().unwrap().into_inner().freeze();
-        let total_size = bytes.len();
+        let total_size = bytes.len() as u64;
         let path = Path::from(path);
         let payload = object_store::PutPayload::from_bytes(bytes);
         store
